+++ examples/trailing-mouse-pointer/src/cljs/example/core.cljs
@@ -17,12 +17,12 @@
   (sb/send-mult-cmd
     switchboard   ;; Below, we're interacting with the switchboard component we just created above.
     [;; First of all, we instantiate and wire a couple fo different components.
-     [:cmd/wire-comp (sente/component :client/ws-cmp)]        ; WebSocket communication component
-     [:cmd/wire-comp (hist/component :client/histogram-cmp)]  ; UI component for histograms
-     [:cmd/wire-comp (mouse/component :client/mouse-cmp)]     ; UI component for capturing mouse moves
-     [:cmd/wire-comp (store/component :client/store-cmp)]     ; Data store component
-     [:cmd/wire-comp (jvmstats/component :client/jvmstats-cmp "jvm-stats-frame")] ;  UI component: JVM stats
-     [:cmd/wire-comp (obs/component :client/observer-cmp conf/observer-cfg-map)]  ; UI component for observing system
+     [:cmd/init-comp (sente/cmp-map :client/ws-cmp)] ; WebSocket communication component
+     [:cmd/init-comp (hist/cmp-map :client/histogram-cmp)]  ; histograms component
+     [:cmd/init-comp (mouse/cmp-map :client/mouse-cmp)]     ; UI component for capturing mouse moves
+     [:cmd/init-comp (store/cmp-map :client/store-cmp)]     ; Data store component
+     [:cmd/init-comp (jvmstats/cmp-map :client/jvmstats-cmp "jvm-stats-frame")] ;  UI component: JVM stats
+     [:cmd/init-comp (obs/cmp-map :client/observer-cmp conf/observer-cfg-map)]  ; UI component for observing system
 
      ;; Then, messages of a given type are wired from one component to another.
      [:cmd/route-all {:from :client/mouse-cmp :to :client/ws-cmp}]